Refactor: summarization logic into SummarizerService class

+++ controllers/LearningTrailController.js
@@ -1,21 +1,11 @@
 const fs = require('fs');
 const LearningTrail = require('../models/LearningTrail');
-const { ChatOpenAI } = require('@langchain/openai');
-const { PromptTemplate } = require('@langchain/core/prompts');
-const { TokenTextSplitter } = require('langchain/text_splitter');
-const { PDFLoader } = require('@langchain/community/document_loaders/fs/pdf');
-const { z } = require('zod');
-const { StructuredOutputParser } = require('langchain/output_parsers');
+const SummarizerService = require('../services/SummarizerService');
 
 class LearningTrailController {
 
-	_getSummarizationPipeline = async () => {
-		const { pipeline } = await import('@xenova/transformers');
-		return await pipeline('summarization', 'Xenova/distilbart-cnn-6-6');
-	}
-
 	_validateData = (req) => {
-		let errors = [];
+		const errors = [];
 
 		if (!req.body.title) {
 			errors.push({ error: 'Informe o título da trilha', element: 'title' });
@@ -40,41 +30,6 @@ class LearningTrailController {
 		return errors;
 	}
 
-	_summarizeDocuments = async (documents) => {
-		const model = new ChatOpenAI({
-			openAIApiKey: process.env.OPENAI_API_KEY,
-			model: 'gpt-4o-mini',
-			temperature: 0.1
-		});
-
-		const outputParser = StructuredOutputParser.fromZodSchema(
-			z.array(z.object({
-				name: z.string().describe("The name of the section"),
-				content: z.string().describe("The content of the section")
-			}))
-		);
-
-		const prompt = PromptTemplate.fromTemplate(`
-			Write a concise summary of the following text delimited by triple backquotes and group then into sections.
-			The summarized content must keep the submitted text language.
-			Formatting instructions: {formatInstructions}
-			Text: \`\`\`{inputText}\`\`\`
-		`);
-
-		let summarizedDocuments = [];
-		for (let document of documents) {
-			const chain = prompt.pipe(model).pipe(outputParser);
-			const response = await chain.invoke({
-				inputText: document.pageContent,
-				formatInstructions: outputParser.getFormatInstructions()
-			});
-
-			summarizedDocuments = summarizedDocuments.concat(response);
-		}
-
-		return summarizedDocuments;
-	}
-
 	getAll = async (req, res) => {
 		const status = req.query.status;
 		const query = { userId: req.user.id };
@@ -93,40 +48,20 @@ class LearningTrailController {
 			return res.status(400).json({ message: 'Houveram erros de validação', errors: validationErrors });
 		}
 
-		const loader = new PDFLoader(req.file.path, {
-			parsedItemSeparator: ''
-		});
+		const summarizerService = new SummarizerService(req.file.path, req.body.pageStart, req.body.pageEnd);
 
-		const documents = await loader.load();
-		const pageStart = req.body.pageStart;
-		const pageEnd = req.body.pageEnd;
-
-		let combinedText = '';
-		documents.slice(pageStart - 1, pageEnd - 1).forEach(doc => {
-			combinedText += `${doc.pageContent}\n`;
-		});
-
-		fs.unlink(`${req.file.destination}/${req.file.filename}`, err => {});
-
-		if (!combinedText) {
-			return res.status(400).json({ message: 'Houveram erros de validação', errors: [{ error: 'O arquivo informado não possui texto', element: 'file' }] });
+		const summarizerErrors = await summarizerService.validate();
+		if (summarizerErrors.length) {
+			return res.status(400).json({ message: 'Houveram erros de validação', errors: summarizerErrors });
 		}
 
-		const splitter = new TokenTextSplitter({
-			modelName: "o200k_base",
-			chunkSize: 256,
-			// chunkSize: 8192,
-			// chunkOverlap: 256
-			chunkOverlap: 24
-		});
-
-		const tokenizedDocuments = await splitter.createDocuments([combinedText]);
-		const summarizedChunks = await this._summarizeDocuments(tokenizedDocuments);
+		const summarizedSections = await summarizerService.summarize();
+		fs.unlink(`${req.file.destination}/${req.file.filename}`, err => {});
 
 		const newTrail = await LearningTrail.create({
 			title: req.body.title,
 			userId: req.user.id,
-			sections: summarizedChunks.map(section => {
+			sections: summarizedSections.map(section => {
 				return { read: false, content: section.content, name: section.name };
 			})
 		});
